+++ assets/src/components/editing/models/popup/PopupContentModal.tsx
@@ -1,6 +1,6 @@
 import { RichTextEditor } from 'components/content/RichTextEditor';
 import { CommandContext } from 'components/editing/commands/interfaces';
-import { Modal } from 'components/editing/toolbars/Modal';
+import { FullScreenModal } from 'components/editing/toolbars/FullScreenModal';
 import * as ContentModel from 'data/content/model';
 import React from 'react';
 
@@ -12,32 +12,67 @@ interface Props {
 }
 export const PopupContentModal = (props: Props) => {
   const [content, setContent] = React.useState(props.model.content);
+  const [trigger, setTrigger] = React.useState(props.model.trigger);
+
+  const isTriggerMode = (mode: ContentModel.PopupTriggerMode) => mode === trigger;
+
+  const triggerSettings = (
+    <form onSubmit={() => {}} id="popup__trigger_mode">
+      <div className="form-check form-switch">
+        <div className="form-group">
+          <label className="form-check-label">
+            <input
+              type="radio"
+              className="form-check-input"
+              onChange={() => setTrigger('hover')}
+              checked={isTriggerMode('hover')}
+            />
+            <p>Trigger on mouseover</p>
+            <small>
+              Students on desktop computers will see this content when they hover over the popup
+              with their mouse. Students on mobile must press. Good for shorter content.
+            </small>
+          </label>
+        </div>
+        <div className="form-group">
+          <label className="form-check-label">
+            <input
+              type="radio"
+              className="form-check-input"
+              onChange={() => setTrigger('click')}
+              checked={isTriggerMode('click')}
+            />
+            <p>Trigger on click</p>
+            <small>
+              All students must click the popup to see this content. Good for longer content or
+              including audio, videos.
+            </small>
+          </label>
+        </div>
+      </div>
+    </form>
+  );
+
   return (
-    <Modal
+    <FullScreenModal
       onCancel={(_e) => props.onCancel()}
       onDone={(newModel) => props.onDone(Object.assign(props.model, newModel))}
     >
-      <div>
-        <h3 className="mb-2">Content</h3>
-        <p className="mb-4">Enter the content</p>
-        <div
-          className="m-auto"
-          style={{
-            width: 300,
-            height: 200,
-            backgroundImage: `url(${props.model.src})`,
-            backgroundPosition: '50% 50%',
-            backgroundSize: 'contain',
-            backgroundRepeat: 'no-repeat',
-          }}
-        ></div>
-        <RichTextEditor
-          editMode={true}
-          text={content}
-          projectSlug={props.commandContext.projectSlug}
-          onEdit={(content) => setContent(content)}
-        />
+      <div className="row">
+        <div className="col-12">
+          <h3 className="mb-2">Popup Content</h3>
+          <p className="mb-4">Shown to students when the popup is active</p>
+          <div className="popup__modal-content">
+            {triggerSettings}
+            <RichTextEditor
+              editMode={true}
+              text={content}
+              projectSlug={props.commandContext.projectSlug}
+              onEdit={(content) => setContent(content)}
+            />
+          </div>
+        </div>
       </div>
-    </Modal>
+    </FullScreenModal>
   );
 };